remote: split runStoreSocket, doReq into Remote.Client.Core
