+++ hnix-store-remote/hnix-store-remote.cabal
@@ -78,6 +78,7 @@ library
     , System.Nix.Store.Remote
     , System.Nix.Store.Remote.Arbitrary
     , System.Nix.Store.Remote.Client
+    , System.Nix.Store.Remote.Client.Core
     , System.Nix.Store.Remote.Logger
     , System.Nix.Store.Remote.MonadStore
     , System.Nix.Store.Remote.Serialize
